+++ packages/build-info/CHANGELOG.md
@@ -64,6 +64,62 @@
   * dependencies
     * @netlify/framework-info bumped from ^9.8.8 to ^9.8.9
 
+## [7.13.2](https://github.com/netlify/build/compare/build-info-v7.13.1...build-info-v7.13.2) (2024-03-26)
+
+
+### Bug Fixes
+
+* don't poll HTTP on startup to prevent dependency cycle between framework dev server and Netlify Dev ([#5558](https://github.com/netlify/build/issues/5558)) ([0291702](https://github.com/netlify/build/commit/0291702cac745ef0f32d9e651340841582b25de9))
+
+## [7.13.1](https://github.com/netlify/build/compare/build-info-v7.13.0...build-info-v7.13.1) (2024-03-18)
+
+
+### Bug Fixes
+
+* **metrics:** add validation/formatting before sending errors to BugSnag ([#5455](https://github.com/netlify/build/issues/5455)) ([e48532f](https://github.com/netlify/build/commit/e48532fc616eaf4e49e49316e424f13179163294))
+
+## [7.13.0](https://github.com/netlify/build/compare/build-info-v7.12.1...build-info-v7.13.0) (2024-02-16)
+
+
+### Features
+
+* add detection for observable framework ([#5526](https://github.com/netlify/build/issues/5526)) ([e1b7b5e](https://github.com/netlify/build/commit/e1b7b5e4a712fe6f4fccc830bf1542f7a30879a0))
+
+## [7.12.1](https://github.com/netlify/build/compare/build-info-v7.12.0...build-info-v7.12.1) (2024-02-12)
+
+
+### Bug Fixes
+
+* switch to @iarna/toml ([#5508](https://github.com/netlify/build/issues/5508)) ([ba81c85](https://github.com/netlify/build/commit/ba81c85aad902380d24fc1722bc65d4feed24b84))
+
+## [7.12.0](https://github.com/netlify/build/compare/build-info-v7.11.5...build-info-v7.12.0) (2024-01-30)
+
+
+### Features
+
+* add latest nx executors to detection ([#5494](https://github.com/netlify/build/issues/5494)) ([ab39be0](https://github.com/netlify/build/commit/ab39be0c5f3788ddd5c981a502d8783cf5ad9263))
+
+## [7.11.5](https://github.com/netlify/build/compare/build-info-v7.11.4...build-info-v7.11.5) (2024-01-15)
+
+
+### Bug Fixes
+
+* update autodetect Nuxt default dir to `dist` ([#5457](https://github.com/netlify/build/issues/5457)) ([43f910d](https://github.com/netlify/build/commit/43f910dad53767d1068d8b9c92753c4342683dfb))
+
+## [7.11.4](https://github.com/netlify/build/compare/build-info-v7.11.3...build-info-v7.11.4) (2023-12-20)
+
+
+### Bug Fixes
+
+* recommend ng17 dist dir based on builder ([#5445](https://github.com/netlify/build/issues/5445)) ([494ffcc](https://github.com/netlify/build/commit/494ffcc04b8038a437c1c5dd4c54ec56d4af79c7))
+
+## [7.11.3](https://github.com/netlify/build/compare/build-info-v7.11.2...build-info-v7.11.3) (2023-12-06)
+
+
+### Bug Fixes
+
+* add clearPublishDirectory to getSettings ([#5433](https://github.com/netlify/build/issues/5433)) ([9770391](https://github.com/netlify/build/commit/9770391b5b73edfde3ea8405404b008df1b7aa1f))
+
 ## [7.11.2](https://github.com/netlify/build/compare/build-info-v7.11.1...build-info-v7.11.2) (2023-12-05)