+++ packages/opentelemetry-sdk-setup/CHANGELOG.md
@@ -1,5 +1,68 @@
 # Changelog
 
+## [1.1.0](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.5...opentelemetry-sdk-setup-v1.1.0) (2024-03-28)
+
+
+### Features
+
+* instrument child process with open telemetry ([#5556](https://github.com/netlify/build/issues/5556)) ([e93feb4](https://github.com/netlify/build/commit/e93feb4fe6510fd92abd94652ecce0bebfd31ab1))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/opentelemetry-utils bumped from ~1.0.3 to ~1.1.0
+
+## [1.0.5](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.4...opentelemetry-sdk-setup-v1.0.5) (2024-03-14)
+
+
+### Bug Fixes
+
+* **otel:** make the otel-api package a peer dep of our build ecosystem ([#5547](https://github.com/netlify/build/issues/5547)) ([a24f68f](https://github.com/netlify/build/commit/a24f68fc67f4f6d464e0f86c2309debb70d22cd9))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/opentelemetry-utils bumped from ^1.0.2 to ^1.0.3
+
+## [1.0.4](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.3...opentelemetry-sdk-setup-v1.0.4) (2024-03-11)
+
+
+### Bug Fixes
+
+* **deps:** bump honeycomb-otel ([#5542](https://github.com/netlify/build/issues/5542)) ([239de6f](https://github.com/netlify/build/commit/239de6fa1178ad3d906155b8edf7c96c03edc3cd))
+
+## [1.0.3](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.2...opentelemetry-sdk-setup-v1.0.3) (2024-02-19)
+
+
+### Bug Fixes
+
+* **deps:** update opentelemetry-js monorepo ([#5528](https://github.com/netlify/build/issues/5528)) ([d6cc768](https://github.com/netlify/build/commit/d6cc7681af3ec78a434650d96dc051beebfd7a44))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/opentelemetry-utils bumped from ^1.0.1 to ^1.0.2
+
+## [1.0.2](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.1...opentelemetry-sdk-setup-v1.0.2) (2024-01-08)
+
+
+### Bug Fixes
+
+* **tracing:** cleanup otel setup ([#5446](https://github.com/netlify/build/issues/5446)) ([2d10996](https://github.com/netlify/build/commit/2d10996ee91e7640be0e4b53105ae30cb42752fe))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/opentelemetry-utils bumped from ^1.0.0 to ^1.0.1
+
 ## [1.0.1](https://github.com/netlify/build/compare/opentelemetry-sdk-setup-v1.0.0...opentelemetry-sdk-setup-v1.0.1) (2023-11-23)